Timestamps format modification: time and baseTime uses Float instead of Long. PoC testing.

+++ leshan-client-core/src/test/java/org/eclipse/leshan/client/send/ManualDataSenderTest.java
@@ -86,7 +86,7 @@ public class ManualDataSenderTest {
 
         // ensure that sent values equals collected ones.
         TimestampedLwM2mNodes lastValuesSent = fakeDataSenderManager.getLastValuesSent();
-        List<Long> timestamps = new ArrayList<>(lastValuesSent.getTimestamps());
+        List<Float> timestamps = new ArrayList<>(lastValuesSent.getTimestamps());
         Assert.assertEquals(3, timestamps.size());
         Assert.assertEquals(firstValue, lastValuesSent.getNodesAt(timestamps.get(0)));
         Assert.assertEquals(secondValue, lastValuesSent.getNodesAt(timestamps.get(1)));
