+++ leshan-core/src/test/java/org/eclipse/leshan/core/node/TimestampedLwM2mNodesTest.java
@@ -37,7 +37,7 @@ public class TimestampedLwM2mNodesTest {
         TimestampedLwM2mNodes tsNodes = getExampleTimestampedLwM2mNodes();
 
         // when
-        Map<LwM2mPath, LwM2mNode> tsNodesMap = tsNodes.getNodesAt(123L);
+        Map<LwM2mPath, LwM2mNode> tsNodesMap = tsNodes.getNodesAt(123f);
 
         // then
         assertNotNull(tsNodesMap);
@@ -65,7 +65,7 @@ public class TimestampedLwM2mNodesTest {
         TimestampedLwM2mNodes tsNodes = getExampleTimestampedLwM2mNodes();
 
         // when
-        Map<LwM2mPath, LwM2mNode> tsNodesMap = tsNodes.getNodesAt(0L);
+        Map<LwM2mPath, LwM2mNode> tsNodesMap = tsNodes.getNodesAt(0f);
 
         // then
         assertNull(tsNodesMap);
@@ -107,11 +107,11 @@ public class TimestampedLwM2mNodesTest {
         TimestampedLwM2mNodes tsNodes = getExampleTimestampedLwM2mNodes();
 
         // when
-        Set<Long> timestamps = tsNodes.getTimestamps();
+        Set<Float> timestamps = tsNodes.getTimestamps();
 
         // then
         assertNotNull(timestamps);
-        Iterator<Long> iterator = timestamps.iterator();
+        Iterator<Float> iterator = timestamps.iterator();
         assertEquals(123L, iterator.next().longValue());
         assertEquals(456L, iterator.next().longValue());
     }
@@ -122,11 +122,11 @@ public class TimestampedLwM2mNodesTest {
         TimestampedLwM2mNodes tsNodes = getExampleMixedTimestampLwM2mNodes();
 
         // when
-        Set<Long> timestamps = tsNodes.getTimestamps();
+        Set<Float> timestamps = tsNodes.getTimestamps();
 
         // then
         assertNotNull(timestamps);
-        Iterator<Long> iterator = timestamps.iterator();
+        Iterator<Float> iterator = timestamps.iterator();
         assertEquals(123L, iterator.next().longValue());
         assertNull(iterator.next());
     }
@@ -150,7 +150,7 @@ public class TimestampedLwM2mNodesTest {
         TimestampedLwM2mNodes tsNodes = getExampleTimestampedLwM2mNodes();
 
         // when
-        Set<Long> timestamps = tsNodes.getTimestamps();
+        Set<Float> timestamps = tsNodes.getTimestamps();
 
         // then
         assertNotNull(timestamps);
@@ -161,9 +161,9 @@ public class TimestampedLwM2mNodesTest {
     public void should_raise_exception_for_duplicates() {
         // given
         TimestampedLwM2mNodes.Builder builder = TimestampedLwM2mNodes.builder();
-        builder.put(456L, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
-        builder.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
-        builder.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 112L));
+        builder.put(456f, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
+        builder.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
+        builder.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 112L));
 
         // when
         assertThrows(IllegalArgumentException.class, () -> builder.build());
@@ -171,14 +171,14 @@ public class TimestampedLwM2mNodesTest {
 
     @Test
     public void should_raise_exception_id_path_does_not_match_node() {
-        TimestampedLwM2mNodes.Builder builder = TimestampedLwM2mNodes.builder().put(456L, new LwM2mPath("/0/0/2"),
+        TimestampedLwM2mNodes.Builder builder = TimestampedLwM2mNodes.builder().put(456f, new LwM2mPath("/0/0/2"),
                 LwM2mResourceInstance.newIntegerInstance(0, 222L));
         assertThrows(IllegalArgumentException.class, () -> builder.build());
     }
 
     @Test
     public void should_raise_exception_path_does_not_match_id() {
-        TimestampedLwM2mNodes.Builder builder = TimestampedLwM2mNodes.builder().put(456L, new LwM2mPath("/0/0/2"),
+        TimestampedLwM2mNodes.Builder builder = TimestampedLwM2mNodes.builder().put(456f, new LwM2mPath("/0/0/2"),
                 LwM2mSingleResource.newIntegerResource(1, 222L));
         assertThrows(IllegalArgumentException.class, () -> builder.build());
     }
@@ -187,9 +187,9 @@ public class TimestampedLwM2mNodesTest {
     public void should_not_raise_exception_for_duplicates() {
         // given
         TimestampedLwM2mNodes.Builder builder = TimestampedLwM2mNodes.builder().raiseExceptionOnDuplicate(false);
-        builder.put(456L, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
-        builder.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
-        builder.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 112L));
+        builder.put(456f, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
+        builder.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
+        builder.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 112L));
 
         // when
         TimestampedLwM2mNodes nodes = builder.build();
@@ -197,29 +197,29 @@ public class TimestampedLwM2mNodesTest {
         // then
         assertNotNull(nodes);
         TimestampedLwM2mNodes.Builder expected = TimestampedLwM2mNodes.builder();
-        expected.put(456L, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
-        expected.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 112L));
+        expected.put(456f, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
+        expected.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 112L));
         assertEquals(expected.build(), nodes);
     }
 
     private TimestampedLwM2mNodes getExampleTimestampedLwM2mNodes() {
         TimestampedLwM2mNodes.Builder tsNodes = TimestampedLwM2mNodes.builder();
-        tsNodes.put(456L, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
-        tsNodes.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
+        tsNodes.put(456f, new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
+        tsNodes.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
         return tsNodes.build();
     }
 
     private TimestampedLwM2mNodes getSamePathTimestampedLwM2mNodes() {
         TimestampedLwM2mNodes.Builder tsNodes = TimestampedLwM2mNodes.builder();
-        tsNodes.put(456L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 222L));
-        tsNodes.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
+        tsNodes.put(456f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 222L));
+        tsNodes.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
         return tsNodes.build();
     }
 
     private TimestampedLwM2mNodes getExampleMixedTimestampLwM2mNodes() {
         TimestampedLwM2mNodes.Builder tsNodes = TimestampedLwM2mNodes.builder();
         tsNodes.put(new LwM2mPath("/0/0/2"), LwM2mSingleResource.newIntegerResource(2, 222L));
-        tsNodes.put(123L, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
+        tsNodes.put(123f, new LwM2mPath("/0/0/1"), LwM2mSingleResource.newIntegerResource(1, 111L));
         return tsNodes.build();
     }
 }
